Refactoring the TAF structure
