+++ src/main/java/API/response/StarshipResponse.java
@@ -1,6 +1,6 @@
-package response;
+package API.response;
 
-import Deserializer.DateDeserializer;
+import API.Deserializer.DateDeserializer;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
